// frontend/src/types.ts
export interface AuthTokens {
  access: string;
  refresh: string;
}